tts/app.py
import grpc
from concurrent import futures
import time

# Import generated gRPC files
# e.g., from generated import commentary_to_tts_pb2, commentary_to_tts_pb2_grpc

# Placeholder for actual imports until protos are compiled
class Placeholder_TtsServicer:
    def SynthesizeSpeech(self, request, context):
        print(f"TTS Service: Received text (Req ID: {request.request_id}): '{request.text}' for lang: {request.language_code}")

        # Dummy TTS logic
        # In a real service, this would call a TTS engine
        dummy_audio_bytes = b"\x00\x01\x02\x03\x04" # Tiny placeholder for audio data
        audio_format = "dummy_pcm"

        print(f"TTS Service: Synthesized dummy audio for Req ID: {request.request_id}. Format: {audio_format}, Length: {len(dummy_audio_bytes)}")

        # return commentary_to_tts_pb2.SpeechSynthesisResponse(
        #     status="COMPLETED_DUMMY",
        #     audio_content=dummy_audio_bytes,
        #     audio_format=audio_format,
        #     request_id=request.request_id
        # )
        # Placeholder return until protos are compiled.
        class DummySpeechResponse:
            def __init__(self, status, audio_content, audio_format, request_id):
                self.status = status
                self.audio_content = audio_content
                self.audio_format = audio_format
                self.request_id = request_id
        return DummySpeechResponse(status="COMPLETED_DUMMY", audio_content=dummy_audio_bytes, audio_format=audio_format, request_id=request.request_id)

def serve_tts():
    # server = grpc.server(futures.ThreadPoolExecutor(max_workers=10))
    # commentary_to_tts_pb2_grpc.add_CommentaryToTtsServicer_to_server(Placeholder_TtsServicer(), server)
    # server.add_insecure_port('[::]:50052')
    # print("TTS Service listening on port 50052...")
    # server.start()
    # server.wait_for_termination()
    print("TTS Service: Placeholder serve function. Run 'generate_protos.sh' first.")
    print("Then uncomment the gRPC server code and imports.")
    # This part is just to keep the container running for the dummy setup
    print("Starting dummy TTS server loop (does nothing without gRPC setup)...")
    try:
        while True:
            time.sleep(60)
            print("TTS service still alive...")
    except KeyboardInterrupt:
        print("TTS service stopping.")

if __name__ == '__main__':
    serve_tts() 